requirements-lock.txt
streamlit==1.55.0
pandas==2.3.3
pyyaml==6.0.3
pytest==9.0.2
ruff==0.14.11
mypy==1.19.1
